Very basic config
This is a hard-coded configuration with one subnet, a single security group,
and also has a hard-coded resource group rather than on that is managed
elsewhere in a more convenient location.

This isn't good for long-term use, it's just a starting point.

Be sure to update versions.tf to include the subscription ID.