Build out hub/spoke network
This turns the basic one-off network config into a hub/spoke network
design. This features a central hub network, which has a subnet configured
for a firewall, and the related firewall config (all delagated to a its
own config file.

As well, this sets up 2 spoke networks with peering connections to the
hub network. As well, there are routes to the firewall. This forces
traffic through the firewall, which also handles NAT.