Dynamic hub/spoke design
This creates a hub network, and an arbitrary number of spoke networks, which
can be generated by defining a mapping of network names, CIDR bocks, the
resource group they belong to, and even a list of subnets that you want to
be created in each spoke network.

The subnets are generated and routes are created in them that route traffic
to the hub network and through the firewall. Note that this contains a default
value for the network variable that spins up sample networks.